Fix Bugs - Image Clamp Range - MemCpy - Assign

+++ OnnxLibrary/SingingVoiceConversion/Modules/src/InferTools/TensorExtractor/BaseTensorExtractor.cpp
@@ -199,37 +199,37 @@ DragonianLibSTL::Vector<float> LibSvcTensorExtractor::InterpUVF0(
 {
 	if (PaddedIndex == size_t(-1))
 		PaddedIndex = F0.Size();
-	DragonianLibSTL::Vector<double> NUVF0;
-	DragonianLibSTL::Vector<double> UVF0Indices, NUVF0Indices;
+	DragonianLibSTL::Vector<float> NUVF0;
+	DragonianLibSTL::Vector<float> UVF0Indices, NUVF0Indices;
 	UVF0Indices.Reserve(F0.Size());
 	NUVF0.Reserve(F0.Size());
 	NUVF0Indices.Reserve(F0.Size());
 	if (F0[0] < 0.0001f)
 	{
-		NUVF0.EmplaceBack(0);
-		NUVF0Indices.EmplaceBack(0);
+		NUVF0.EmplaceBack(0.f);
+		NUVF0Indices.EmplaceBack(0.f);
 	}
 	for (size_t i = 1; i < PaddedIndex; ++i)
 	{
 		if (F0[i] < 0.0001f)
-			UVF0Indices.EmplaceBack((double)i);
+			UVF0Indices.EmplaceBack((float)i);
 		else
 		{
-			NUVF0.EmplaceBack((double)F0[i]);
-			NUVF0Indices.EmplaceBack((double)i);
+			NUVF0.EmplaceBack(F0[i]);
+			NUVF0Indices.EmplaceBack((float)i);
 		}
 	}
 	if (UVF0Indices.Empty() || NUVF0Indices.Empty())
 		return F0;
 
-	NUVF0Indices.EmplaceBack(F0.Size());
-	NUVF0.EmplaceBack(0.);
-	DragonianLibSTL::Vector<double> UVF0(F0.Size());
+	NUVF0Indices.EmplaceBack((float)F0.Size());
+	NUVF0.EmplaceBack(0.f);
+	DragonianLibSTL::Vector<float> UVF0(F0.Size());
 	DragonianLibSTL::Vector<float> Of0 = F0;
 	//interp1(NUVF0Indices.Data(), NUVF0.Data(), (int)NUVF0.Size(),
 	//	UVF0Indices.Data(), (int)UVF0Indices.Size(), UVF0.Data());
 	for (size_t i = 0; i < UVF0Indices.Size(); ++i)
-		Of0[size_t(UVF0Indices[i])] = (float)UVF0[i];
+		Of0[size_t(UVF0Indices[i])] = UVF0[i];
 	return Of0;
 }
